clip to screen

+++ source/Chart/Chart.h
@@ -15,9 +15,9 @@ public:
     ~Chart();
     juce::String name;
     juce::MemoryBlock midiData;
-    std::multimap<long, ChartEvent> events;
+    std::multimap<long long, ChartEvent> events;
 
-    long countInTime = 2000;
+    long long countInTime = 2000;
 };
 
 #endif //UPBEAT_CHART_H
